[#118] Feat: useAuth에 로그인, 회원가입 api 연동 함수 추가
